+++ trestle/core/catalog/catalog_interface.py
@@ -13,24 +13,17 @@
 # limitations under the License.
 """Provide interface to catalog allowing queries and operations at control level."""
 
-import copy
 import logging
 import pathlib
 from dataclasses import dataclass
 from typing import Any, Callable, Dict, Iterator, List, Optional, Set, Tuple
 
 import trestle.common.const as const
-import trestle.core.generators as gens
-import trestle.core.generic_oscal as generic
 import trestle.oscal.catalog as cat
 from trestle.common.err import TrestleError
-from trestle.common.file_utils import prune_empty_dirs
 from trestle.common.list_utils import as_dict, as_filtered_list, as_list, delete_item_from_list, deep_set, get_item_from_list, none_if_empty, set_or_pop  # noqa E501
-from trestle.common.model_utils import ModelUtils
-from trestle.core.control_context import ContextPurpose, ControlContext
-from trestle.core.control_interface import CompDict, ComponentImpInfo, ControlInterface, ParameterRep
-from trestle.core.control_reader import ControlReader
-from trestle.core.control_writer import ControlWriter
+from trestle.core.control_context import ControlContext
+from trestle.core.control_interface import CompDict, ComponentImpInfo, ControlInterface
 from trestle.oscal import common
 from trestle.oscal import component as comp
 from trestle.oscal import profile as prof
@@ -677,25 +670,6 @@ class CatalogInterface():
                 return prop.value, ns
         return None, None
 
-    def _write_control_into_dir(
-        self,
-        context: ControlContext,
-        control: cat.Control,
-        part_id_map: Dict[str, Dict[str, str]],
-        found_control_alters: List[prof.Alter]
-    ):
-        # we need to create the dir structure on demand because we don't know a priori what groups are included
-        _, group_title, _ = self.get_group_info_by_control(control.id)
-        group_dir = context.md_root
-        control_path = self.get_control_path(control.id)
-        for sub_dir in control_path:
-            group_dir = group_dir / sub_dir
-            if not group_dir.exists():
-                group_dir.mkdir(parents=True, exist_ok=True)
-
-        writer = ControlWriter()
-        writer.write_control_for_editing(context, control, group_dir, group_title, part_id_map, found_control_alters)
-
     # FIXME typing
     @staticmethod
     def _get_all_rules_params_and_vals(context: ControlContext) -> None:
@@ -715,99 +689,6 @@ class CatalogInterface():
     def _prune_controls(md_path: pathlib.Path, written_controls: Set[str]):
         """Search directory and remove any controls that were not written out."""
         pass
-
-    def write_catalog_as_profile_markdown(
-        self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]
-    ) -> None:
-        """Write out the catalog as profile markdown."""
-        # Get the list of params for this profile from its set_params
-        # this is just from the set_params
-        full_param_dict = CatalogInterface._get_full_profile_param_dict(context.profile)
-
-        label_map = self.get_statement_part_id_map(True)
-        found_alters, _, _ = CatalogInterface.read_additional_content(
-            context.md_root, context.required_sections, label_map, context.sections_dict, context.to_markdown
-        )
-
-        # write out the controls
-        for control in self.get_all_controls_from_catalog(True):
-            # here we do special handling of how set-parameters merge with the yaml header
-            new_context = ControlContext.clone(context)
-            new_context.merged_header = {}
-            if new_context.inherited_props:
-                inherited_props = new_context.inherited_props.get(control.id, None)
-                if inherited_props:
-                    # build set in order of list so that duplicates will have final value stick, then convert to list
-                    unique_props = list({prop['name']: prop for prop in inherited_props}.values())
-                    new_context.merged_header[const.TRESTLE_INHERITED_PROPS_TAG] = unique_props
-            # get all params and vals for this control from the resolved profile catalog with block adds in effect
-            control_param_dict = ControlInterface.get_control_param_dict(control, False)
-            set_param_dict: Dict[str, str] = {}
-            for param_id, param_dict in control_param_dict.items():
-                # if the param is in the full_param_dict, load its contents first and mark as profile-values
-                display_name = ''
-                if param_id in full_param_dict:
-                    # get the param from the profile set_param
-                    param = full_param_dict[param_id]
-                    display_name, _ = CatalogInterface._get_display_name_and_ns(param)
-                    # assign its contents to the dict
-                    new_dict = ModelUtils.parameter_to_dict(param, True)
-                    if const.VALUES in new_dict:
-                        new_dict[const.PROFILE_VALUES] = new_dict[const.VALUES]
-                        new_dict.pop(const.VALUES)
-                    # then insert the original, incoming values as values
-                    if param_id in control_param_dict:
-                        orig_param = control_param_dict[param_id]
-                        orig_dict = ModelUtils.parameter_to_dict(orig_param, True)
-                        # pull only the values from the actual control dict
-                        # all the other elements are from the profile set_param
-                        new_dict[const.VALUES] = orig_dict.get(const.VALUES, None)
-                else:
-                    # if the profile doesnt change this param at all, show it in the header with values
-                    tmp_dict = ModelUtils.parameter_to_dict(param_dict, True)
-                    values = tmp_dict.get('values', None)
-                    new_dict = {'id': param_id, 'values': values}
-                new_dict.pop('id', None)
-                if display_name:
-                    new_dict[const.DISPLAY_NAME] = display_name
-                set_param_dict[param_id] = new_dict
-            if set_param_dict:
-                if const.SET_PARAMS_TAG not in new_context.yaml_header:
-                    new_context.yaml_header[const.SET_PARAMS_TAG] = {}
-                # FIXME is this intended behavior
-                if new_context.overwrite_header_values:
-                    # update the control params with new values
-                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
-                        if key in control_param_dict:
-                            set_param_dict[key] = value
-                else:
-                    # update the control params with any values in yaml header not set in control
-                    # need to maintain order in the set_param_dict
-                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
-                        if key in control_param_dict and key not in set_param_dict:
-                            set_param_dict[key] = value
-                new_context.merged_header[const.SET_PARAMS_TAG] = set_param_dict
-            elif const.SET_PARAMS_TAG in new_context.merged_header:
-                # need to cull any params that are not in control
-                pop_list: List[str] = []
-                for key in new_context.merged_header[const.SET_PARAMS_TAG].keys():
-                    if key not in control_param_dict:
-                        pop_list.append(key)
-                for pop in pop_list:
-                    new_context.merged_header[const.SET_PARAMS_TAG].pop(pop)
-
-            found_control_alters = [alter for alter in found_alters if alter.control_id == control.id]
-
-            self._write_control_into_dir(new_context, control, part_id_map, found_control_alters)
-
-    @staticmethod
-    def _read_comp_info_from_md(control_file_path: pathlib.Path,
-                                context: ControlContext) -> Tuple[Dict[str, Any], CompDict]:
-        comp_dict = {}
-        md_header = {}
-        if control_file_path.exists():
-            md_header = ControlReader.read_control_info_from_md(control_file_path, comp_dict, context)
-        return md_header, comp_dict
 
     def _get_control_memory_info(self, control_id: str, context) -> Dict[str, Any]:
         # set COMP_DEF_RULES_TAG, RULES_PARAMS_TAG, COMP_DEF_RULES_PARAM_VALS_TAG, SET_PARAMS_TAG
@@ -849,269 +730,6 @@ class CatalogInterface():
 
         return header, context.comp_dict
 
-    def _merge_header_and_comp_dict(
-        self, control: cat.Control, control_file_path: pathlib.Path, context: ControlContext
-    ) -> None:
-        """
-        Merge the header and the comp_dict.
-
-        x-trestle-comp-def-rules:
-        - name: XCCDF
-            description: The XCCDF must be compliant
-        x-trestle-rules-params:
-        - name: foo_length
-            description: minimum_foo_length
-            rule-id: XCCDF
-            options: '["6", "9"]'
-        x-trestle-comp-def-param-vals:
-        quantity_available: '500'
-        foo_length: '6'
-        x-trestle-global:
-        profile-title: NIST Special Publication 800-53 Revision 5 MODERATE IMPACT BASELINE
-        x-trestle-set-params:
-        ac-1_prm_1:
-            values: Param_1_value_in_catalog
-        ac-1_prm_2:
-            values:
-        ac-1_prm_3:
-            values:
-        ac-1_prm_4:
-            values:
-        ac-1_prm_5:
-            values:
-        ac-1_prm_6:
-            values:
-        ac-1_prm_7:
-            values:
-        ---
-
-        # now have all rules in context.rules_dict and all rules_params in context.rules_params_dict
-        # all set-params per component for each control are in the cat interface
-        # all comp-infos by control and part are in the cat interface
-        #
-        # can now write out catalog and pull from the markdown:
-        # header for param values to set during assem
-        # prose and status for This System
-        # status for all parts that still have rules
-
-        set COMP_DEF_RULES_TAG, RULES_PARAMS_TAG, COMP_DEF_RULES_PARAM_VALS_TAG, SET_PARAMS_TAG
-
-        """
-        memory_header, memory_comp_dict = self._get_control_memory_info(control.id, context)
-        md_header, md_comp_dict = self._read_comp_info_from_md(control_file_path, context)
-        # get select info from md and update the memory comp_dict
-        if const.SSP_MAIN_COMP_NAME in md_comp_dict:
-            sys_memory_dict = memory_comp_dict.get(const.SSP_MAIN_COMP_NAME, {})
-            sys_md_dict = md_comp_dict[const.SSP_MAIN_COMP_NAME]
-            # for This System the md completely replaces what is in memory
-            # but only for items that are in memory, which means they have rules
-            for label, comp_info in memory_comp_dict.items():
-                if label in sys_md_dict:
-                    sys_memory_dict[label] = comp_info
-            md_comp_dict.pop(const.SSP_MAIN_COMP_NAME)
-        for comp_name, md_label_dict in md_comp_dict.items():
-            memory_label_dict = memory_comp_dict.get(comp_name, None)
-            if not memory_label_dict:
-                continue
-            for label, comp_info in md_label_dict.items():
-                if label in memory_label_dict:
-                    memory_label_dict[label].status = comp_info.status
-
-        memory_rules_param_vals = memory_header.get(const.COMP_DEF_RULES_PARAM_VALS_TAG, {})
-        md_rules_param_vals = md_header.get(const.COMP_DEF_RULES_PARAM_VALS_TAG, {})
-        ControlInterface.merge_dicts_deep(memory_rules_param_vals, md_rules_param_vals, True)
-        set_or_pop(memory_header, const.COMP_DEF_RULES_PARAM_VALS_TAG, memory_rules_param_vals)
-        context.merged_header = memory_header
-
-    def write_catalog_as_ssp_markdown(self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]) -> None:
-        """
-        Write out the catalog as component markdown.
-
-        In memory:
-        for each compdef:
-            for each comp:
-                load top level rules
-                for each control_imp:
-                    load set-params
-                    for each imp_req (bound to 1 control):
-                        load set-params
-                        load control level rules and status
-                        load part level rules and status
-                        add as compinfo to control comp_dict
-
-        """
-        # first load all information from memory for all comps and all controls
-        context.rules_dict = {}
-        context.rules_params_dict = {}
-        for comp_def_name in context.comp_def_name_list:
-            context.comp_def, _ = ModelUtils.load_top_level_model(
-                context.trestle_root,
-                comp_def_name,
-                comp.ComponentDefinition
-            )
-            for component in as_list(context.comp_def.components):
-                context.component = component
-                context.comp_name = component.title
-                # get top level rule info applying to all controls
-                comp_rules_dict, comp_rules_params_dict = ControlInterface.get_rules_and_params_dict_from_item(component)  # noqa E501
-                context.rules_dict.update(comp_rules_dict)
-                context.rules_params_dict.update(comp_rules_params_dict)
-                for control_imp in as_list(component.control_implementations):
-                    control_imp_rules_dict, control_imp_rules_params_dict = ControlInterface.get_rules_and_params_dict_from_item(control_imp)  # noqa E501
-                    context.rules_dict.update(control_imp_rules_dict)
-                    context.rules_params_dict.update(control_imp_rules_params_dict)
-                    ci_set_params = ControlInterface.get_set_params_from_item(control_imp)
-                    for imp_req in as_list(control_imp.implemented_requirements):
-                        control_part_id_map = part_id_map.get(imp_req.control_id, {})
-                        control_rules = ControlInterface.get_rule_list_for_item(imp_req)
-                        if control_rules:
-                            status = ControlInterface.get_status_from_props(imp_req)
-                            comp_info = ComponentImpInfo(imp_req.description, control_rules, status)
-                            self.add_comp_info(imp_req.control_id, context.comp_name, '', comp_info)
-                            set_params = copy.deepcopy(ci_set_params)
-                            set_params.update(ControlInterface.get_set_params_from_item(imp_req))
-                            for set_param in set_params.values():
-                                self.add_comp_set_param(imp_req.control_id, context.comp_name, set_param)
-                            for statement in as_list(imp_req.statements):
-                                rule_list = ControlInterface.get_rule_list_for_item(statement)
-                                if rule_list:
-                                    status = ControlInterface.get_status_from_props(statement)
-                                    if statement.statement_id not in control_part_id_map:
-                                        label = statement.statement_id
-                                        logger.warning(
-                                            f'No statement label found for statement id {label}.  Defaulting to {label}.'  # noqa E501
-                                        )
-                                    else:
-                                        label = control_part_id_map[statement.statement_id]
-                                    comp_info = ComponentImpInfo(statement.description, rule_list, status)
-                                    self.add_comp_info(imp_req.control_id, context.comp_name, label, comp_info)
-
-        # now have all rules in context.rules_dict and all rules_params in context.rules_params_dict
-        # all set-params per component for each control are in the cat interface
-        # all comp-infos by control and part are in the cat interface
-        #
-        # can now write out catalog and pull from the markdown:
-        # header for param values to set during assem
-        # prose and status for This System
-        # status for all parts that still have rules
-
-        for control_id, _ in self._control_comp_dicts.items():
-            control_file_path = self.get_control_file_path(context.md_root, control_id)
-            if not control_file_path:
-                logger.warning(f'Control {control_id} referenced by component is not in the resolved catalog.')
-                continue
-            control_file_path.parent.mkdir(exist_ok=True, parents=True)
-            control = self.get_control(control_id)
-            _, group_title, _ = self.get_group_info_by_control(control_id)
-            # merge the md_header and md_comp_dict with info in cat_interface for this control
-            self._merge_header_and_comp_dict(control, control_file_path, context)
-            control_writer = ControlWriter()
-            control_writer.write_control_for_editing(
-                context, control, control_file_path.parent, group_title, part_id_map, []
-            )
-
-    def write_catalog_as_component_markdown(
-        self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]
-    ) -> None:
-        """Write out the catalog as component markdown."""
-
-        def _update_values(set_param: comp.SetParameter) -> None:
-            if set_param.param_id in control_param_dict:
-                control_param_dict[set_param.param_id].values = set_param.values
-            if set_param.param_id in new_context.rules_param_vals:
-                values = ', '.join([v.__root__ for v in as_list(set_param.values)])
-                new_context.rules_param_vals[set_param.param_id] = values
-
-        control_ids_in_comp_imp = [
-            imp_req.control_id for imp_req in as_list(context.control_implementation.implemented_requirements)
-        ]
-
-        for control in self.get_all_controls_from_catalog(True):
-            if control.id in control_ids_in_comp_imp:
-                new_context = ControlContext.clone(context)
-                control_param_dict = ControlInterface.get_control_param_dict(control, False)
-                for set_param in as_list(new_context.control_implementation.set_parameters):
-                    _update_values(set_param)
-                for imp_req in as_list(new_context.control_implementation.implemented_requirements):
-                    if imp_req.control_id == control.id:
-                        for set_param in as_list(imp_req.set_parameters):
-                            _update_values(set_param)
-
-                if control_param_dict:
-                    new_context.yaml_header[const.PARAM_VALUES_TAG] = {}
-                    for key, param in control_param_dict.items():
-                        new_context.yaml_header[const.PARAM_VALUES_TAG][key] = none_if_empty(
-                            ControlInterface.param_to_str(param, ParameterRep.VALUE_OR_EMPTY_STRING)
-                        )
-
-                self._write_control_into_dir(new_context, control, part_id_map, [])
-
-    def write_catalog_as_catalog(self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]) -> None:
-        """Write the catalog as a simple catalog."""
-        # write out the controls
-        for control in self.get_all_controls_from_catalog(True):
-            # here we do special handling of how set-parameters merge with the yaml header
-            new_context = ControlContext.clone(context)
-
-            control_param_dict = ControlInterface.get_control_param_dict(control, False)
-            set_param_dict: Dict[str, str] = {}
-            for param_id, param_dict in control_param_dict.items():
-                tmp_dict = ModelUtils.parameter_to_dict(param_dict, True)
-                values = tmp_dict.get('values', None)
-                new_dict = {'values': values}
-                set_param_dict[param_id] = new_dict
-            if set_param_dict:
-                if const.SET_PARAMS_TAG not in new_context.yaml_header:
-                    new_context.yaml_header[const.SET_PARAMS_TAG] = {}
-                if new_context.overwrite_header_values:
-                    # update the control params with new values
-                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
-                        if key in control_param_dict:
-                            set_param_dict[key] = value
-                else:
-                    # update the control params with any values in yaml header not set in control
-                    # need to maintain order in the set_param_dict
-                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
-                        if key in control_param_dict and key not in set_param_dict:
-                            set_param_dict[key] = value
-                new_context.yaml_header[const.SET_PARAMS_TAG] = set_param_dict
-            elif const.SET_PARAMS_TAG in new_context.yaml_header:
-                # need to cull any params that are not in control
-                pop_list: List[str] = []
-                for key in new_context.yaml_header[const.SET_PARAMS_TAG].keys():
-                    if key not in control_param_dict:
-                        pop_list.append(key)
-                for pop in pop_list:
-                    new_context.yaml_header[const.SET_PARAMS_TAG].pop(pop)
-
-            self._write_control_into_dir(new_context, control, part_id_map, [])
-
-    def write_catalog_as_markdown(self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]) -> None:
-        """
-        Write out the catalog controls from dict as markdown files to the specified directory.
-
-        Args:
-            context: The context of the catalog markdown creation.
-            part_id_map: Mapping of part_id to label for all controls
-
-        Returns:
-            None
-        """
-        # create the directory in which to write the control markdown files
-        context.md_root.mkdir(exist_ok=True, parents=True)
-
-        if context.purpose == ContextPurpose.PROFILE:
-            self.write_catalog_as_profile_markdown(context, part_id_map)
-        elif context.purpose == ContextPurpose.COMPONENT:
-            self.write_catalog_as_component_markdown(context, part_id_map)
-        elif context.purpose == ContextPurpose.SSP:
-            self.write_catalog_as_ssp_markdown(context, part_id_map)
-        else:
-            self.write_catalog_as_catalog(context, part_id_map)
-
-        # prune any directories that have no markdown files
-        prune_empty_dirs(context.md_root, '*.md')
-
     @staticmethod
     def _get_group_ids_and_dirs(md_path: pathlib.Path) -> Dict[str, pathlib.Path]:
         """
@@ -1130,110 +748,6 @@ class CatalogInterface():
             sorted_id_map[key] = id_map[key]
         return sorted_id_map
 
-    def read_catalog_from_markdown(self, md_path: pathlib.Path, set_parameters_flag: bool) -> cat.Catalog:
-        """
-        Read the groups and catalog controls from the given directory.
-
-        This will overwrite the existing groups and controls in the catalog.
-        """
-        if not self._catalog:
-            self._catalog = gens.generate_sample_model(cat.Catalog)
-        id_map = CatalogInterface._get_group_ids_and_dirs(md_path)
-        groups: List[cat.Group] = []
-        # read each group dir
-        for group_id, group_dir in id_map.items():
-            control_list_raw = []
-            group_title = ''
-            # Need to get group title from at least one control in this directory
-            # All controls in dir should have same group title
-            # Set group title to the first one found and warn if different non-empty title appears
-            # Controls with empty group titles are tolerated but at least one title must be present or warning given
-            # The special group with no name that has the catalog as parent is just a list and has no title
-            for control_path in group_dir.glob('*.md'):
-                control, control_group_title = ControlReader.read_control(control_path, set_parameters_flag)
-                if control_group_title:
-                    if group_title:
-                        if control_group_title != group_title:
-                            logger.warning(
-                                f'Control {control.id} group title {control_group_title} differs from {group_title}'
-                            )
-                    else:
-                        group_title = control_group_title
-                control_list_raw.append(control)
-            control_list = sorted(control_list_raw, key=lambda control: ControlInterface.get_sort_id(control))
-            if group_id:
-                if not group_title:
-                    logger.warning(f'No group title found in controls for group {group_id}')
-                new_group = cat.Group(id=group_id, title=group_title)
-                new_group.controls = none_if_empty(control_list)
-                groups.append(new_group)
-            else:
-                # if the list of controls has no group id it also has no title and is just the controls of the catalog
-                self._catalog.controls = none_if_empty(control_list)
-        self._catalog.groups = none_if_empty(groups)
-        self._create_control_dict()
-        self._catalog.params = none_if_empty(self._catalog.params)
-        return self._catalog
-
-    @staticmethod
-    def read_catalog_imp_reqs(
-        md_path: pathlib.Path, avail_comps: Dict[str, generic.GenericComponent], context: ControlContext
-    ) -> List[generic.GenericImplementedRequirement]:
-        """Read the full set of control implemented requirements from markdown.
-
-        Args:
-            md_path: Path to the markdown control files, with directories for each group
-            avail_comps: Dict mapping component names to known components
-
-        Returns:
-            List of implemented requirements gathered from each control
-
-        Notes:
-            As the controls are read into the catalog the needed components are added if not already available.
-            avail_comps provides the mapping of component name to the actual component.
-            This is only used for ssp via catalog_interface
-        """
-        imp_req_map: Dict[str, generic.GenericImplementedRequirement] = {}
-        for group_path in CatalogInterface._get_group_ids_and_dirs(md_path).values():
-            for control_file in group_path.glob('*.md'):
-                sort_id, imp_req = ControlReader.read_implemented_requirement(control_file, avail_comps, context)
-                imp_req_map[sort_id] = imp_req
-        return [imp_req_map[key] for key in sorted(imp_req_map.keys())]
-
-    @staticmethod
-    def read_additional_content(
-        md_path: pathlib.Path,
-        required_sections_list: List[str],
-        label_map: Dict[str, Dict[str, str]],
-        sections_dict: Dict[str, str],
-        write_mode: bool
-    ) -> Tuple[List[prof.Alter], Dict[str, Any], Dict[str, str]]:
-        """Read all markdown controls and return list of alters plus control param dict and param sort map."""
-        alters_map: Dict[str, prof.Alter] = {}
-        final_param_dict: Dict[str, Any] = {}
-        param_sort_map: Dict[str, str] = {}
-        for group_path in CatalogInterface._get_group_ids_and_dirs(md_path).values():
-            for control_file in group_path.glob('*.md'):
-                sort_id, control_alters, control_param_dict = ControlReader.read_editable_content(
-                    control_file,
-                    required_sections_list,
-                    label_map,
-                    sections_dict,
-                    write_mode
-                )
-                alters_map[sort_id] = control_alters
-                for param_id, param_dict in control_param_dict.items():
-                    # if profile_values are present, overwrite values with them
-                    if const.PROFILE_VALUES in param_dict:
-                        param_dict[const.VALUES] = param_dict.pop(const.PROFILE_VALUES)
-                        final_param_dict[param_id] = param_dict
-                        param_sort_map[param_id] = sort_id
-        new_alters: List[prof.Alter] = []
-        # fill the alters according to the control sorting order
-        for key in sorted(alters_map.keys()):
-            new_alters.extend(alters_map[key])
-        return new_alters, final_param_dict, param_sort_map
-
     def get_sections(self) -> List[str]:
         """Get the available sections by a full index of all controls."""
         return list(
@@ -1244,20 +758,6 @@ class CatalogInterface():
                 if part.name != const.STATEMENT
             }
         )
-
-    @staticmethod
-    def merge_controls(dest: cat.Control, src: cat.Control, replace_params: bool) -> None:
-        """
-        Merge the src control into dest.
-
-        Args:
-            dest: destination control into which content will be added
-            src: source control with new content
-            replace_params: replace the control params with the new ones
-        """
-        ControlInterface.merge_parts(dest, src)
-        if replace_params:
-            dest.params = src.params
 
     def _find_control_in_group(self, group_id: str) -> Tuple[str, ControlHandle]:
         """
@@ -1270,50 +770,3 @@ class CatalogInterface():
             if control_handle.group_id == group_id:
                 return control_id, control_handle
         raise TrestleError(f'No controls found for group {group_id}')
-
-    def merge_catalog(self, catalog: cat.Catalog, replace_params: bool) -> None:
-        """
-        Merge the provided new catalog controls into the original catalog in this catalog interface.
-
-        Args:
-            catalog: catalog containing controls that are merged into the current catalog of the interface
-            replace_params: replace all params in the control with the new ones
-
-        Notes:
-            This is mainly to support the reading of a catalog from markdown.  It allows retention of content such as
-            metadata and backmatter, along with labels and other parameter attributes that aren't in markdown.
-            The list of controls and group structure is specified by the markdown structure - but this doesn't allow
-            controls to contain controls.  Group lists are specified per directory.
-
-            Reading the markdown tells you groups and controls in them - and groups in groups.
-            Controls cannot change groups.  If the control was in the original json, its parts are replaced,
-            including its parameters.  Only values may be specified.  If no value specified, the value is unset in json.
-        """
-        cat_interface = CatalogInterface(catalog)
-        for src in cat_interface.get_all_controls_from_dict():
-            group_id, _, _ = cat_interface.get_group_info_by_control(src.id)
-            dest = self.get_control(src.id)
-            if dest:
-                dest_group, _, _ = self.get_group_info_by_control(dest.id)
-                if dest_group != group_id:
-                    raise TrestleError(f'Markdown for control {src.id} has different group id.')
-                CatalogInterface.merge_controls(dest, src, replace_params)
-                self.replace_control(dest)
-            else:
-                # this is a new control that isn't already in the merge destination
-                # need to add the control knowing its group must already exist
-                # get group info from an arbitrary control already present in group
-                _, control_handle = self._find_control_in_group(group_id)
-                new_control_handle = copy.deepcopy(control_handle)
-                new_control_handle.control = src
-                # add the control and its handle to the param_dict
-                self._control_dict[src.id] = new_control_handle
-
-        # now need to cull any controls that are not in the src catalog
-        handled_ids = set(cat_interface._control_dict.keys())
-        orig_ids = set(self._control_dict.keys())
-        extra_ids = orig_ids.difference(handled_ids)
-        for extra_id in extra_ids:
-            self._control_dict.pop(extra_id)
-
-        self.update_catalog_controls()
