+++ trestle/core/catalog/catalog_writer.py
@@ -1,0 +1,309 @@
+# Copyright (c) 2021 IBM Corp. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Provide interface to catalog allowing queries and operations at control level."""
+
+import copy
+import logging
+from typing import Dict, List
+
+import trestle.common.const as const
+import trestle.oscal.catalog as cat
+from trestle.core.catalog.catalog_interface import CatalogInterface
+from trestle.common.list_utils import as_dict, as_filtered_list, as_list, delete_item_from_list, deep_set, get_item_from_list, none_if_empty, set_or_pop  # noqa E501
+from trestle.common.model_utils import ModelUtils
+from trestle.core.control_context import ControlContext
+from trestle.core.control_interface import ComponentImpInfo, ControlInterface, ParameterRep
+from trestle.core.control_writer import ControlWriter
+from trestle.oscal import component as comp
+from trestle.oscal import profile as prof
+
+logger = logging.getLogger(__name__)
+
+
+class CatalogWriter():
+    """
+    Catalog writer.
+
+    Catalog writer handles all operation related to writing
+    catalog to markdown.
+    """
+
+    def __init__(self, catalog_interface: CatalogInterface):
+        """Initialize catalog writer."""
+        self._catalog_interface = catalog_interface
+
+    def write_catalog_as_profile_markdown(
+        self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]], md_alters: List[prof.Alter]
+    ) -> None:
+        """Write out the catalog as profile markdown."""
+        # Get the list of params for this profile from its set_params
+        # this is just from the set_params
+        full_param_dict = CatalogInterface._get_full_profile_param_dict(context.profile)
+
+        # write out the controls
+        for control in self._catalog_interface.get_all_controls_from_catalog(True):
+            # here we do special handling of how set-parameters merge with the yaml header
+            new_context = ControlContext.clone(context)
+            new_context.merged_header = {}
+            if new_context.inherited_props:
+                inherited_props = new_context.inherited_props.get(control.id, None)
+                if inherited_props:
+                    # build set in order of list so that duplicates will have final value stick, then convert to list
+                    unique_props = list({prop['name']: prop for prop in inherited_props}.values())
+                    new_context.merged_header[const.TRESTLE_INHERITED_PROPS_TAG] = unique_props
+            # get all params and vals for this control from the resolved profile catalog with block adds in effect
+            control_param_dict = ControlInterface.get_control_param_dict(control, False)
+            set_param_dict: Dict[str, str] = {}
+            for param_id, param_dict in control_param_dict.items():
+                # if the param is in the full_param_dict, load its contents first and mark as profile-values
+                display_name = ''
+                if param_id in full_param_dict:
+                    # get the param from the profile set_param
+                    param = full_param_dict[param_id]
+                    display_name, _ = CatalogInterface._get_display_name_and_ns(param)
+                    # assign its contents to the dict
+                    new_dict = ModelUtils.parameter_to_dict(param, True)
+                    if const.VALUES in new_dict:
+                        new_dict[const.PROFILE_VALUES] = new_dict[const.VALUES]
+                        new_dict.pop(const.VALUES)
+                    # then insert the original, incoming values as values
+                    if param_id in control_param_dict:
+                        orig_param = control_param_dict[param_id]
+                        orig_dict = ModelUtils.parameter_to_dict(orig_param, True)
+                        # pull only the values from the actual control dict
+                        # all the other elements are from the profile set_param
+                        new_dict[const.VALUES] = orig_dict.get(const.VALUES, None)
+                else:
+                    # if the profile doesnt change this param at all, show it in the header with values
+                    tmp_dict = ModelUtils.parameter_to_dict(param_dict, True)
+                    values = tmp_dict.get('values', None)
+                    new_dict = {'id': param_id, 'values': values}
+                new_dict.pop('id', None)
+                if display_name:
+                    new_dict[const.DISPLAY_NAME] = display_name
+                set_param_dict[param_id] = new_dict
+            if set_param_dict:
+                if const.SET_PARAMS_TAG not in new_context.yaml_header:
+                    new_context.yaml_header[const.SET_PARAMS_TAG] = {}
+                # FIXME is this intended behavior
+                if new_context.overwrite_header_values:
+                    # update the control params with new values
+                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
+                        if key in control_param_dict:
+                            set_param_dict[key] = value
+                else:
+                    # update the control params with any values in yaml header not set in control
+                    # need to maintain order in the set_param_dict
+                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
+                        if key in control_param_dict and key not in set_param_dict:
+                            set_param_dict[key] = value
+                new_context.merged_header[const.SET_PARAMS_TAG] = set_param_dict
+            elif const.SET_PARAMS_TAG in new_context.merged_header:
+                # need to cull any params that are not in control
+                pop_list: List[str] = []
+                for key in new_context.merged_header[const.SET_PARAMS_TAG].keys():
+                    if key not in control_param_dict:
+                        pop_list.append(key)
+                for pop in pop_list:
+                    new_context.merged_header[const.SET_PARAMS_TAG].pop(pop)
+
+            found_control_alters = [alter for alter in md_alters if alter.control_id == control.id]
+
+            self._write_control_into_dir(new_context, control, part_id_map, found_control_alters)
+
+    def write_catalog_as_ssp_markdown(self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]) -> None:
+        """
+        Write out the catalog as component markdown.
+
+        In memory:
+        for each compdef:
+            for each comp:
+                load top level rules
+                for each control_imp:
+                    load set-params
+                    for each imp_req (bound to 1 control):
+                        load set-params
+                        load control level rules and status
+                        load part level rules and status
+                        add as compinfo to control comp_dict
+
+        """
+        # first load all information from memory for all comps and all controls
+        context.rules_dict = {}
+        context.rules_params_dict = {}
+        for comp_def_name in context.comp_def_name_list:
+            context.comp_def, _ = ModelUtils.load_top_level_model(
+                context.trestle_root,
+                comp_def_name,
+                comp.ComponentDefinition
+            )
+            for component in as_list(context.comp_def.components):
+                context.component = component
+                context.comp_name = component.title
+                # get top level rule info applying to all controls
+                comp_rules_dict, comp_rules_params_dict = ControlInterface.get_rules_and_params_dict_from_item(component)  # noqa E501
+                context.rules_dict.update(comp_rules_dict)
+                context.rules_params_dict.update(comp_rules_params_dict)
+                for control_imp in as_list(component.control_implementations):
+                    control_imp_rules_dict, control_imp_rules_params_dict = ControlInterface.get_rules_and_params_dict_from_item(control_imp)  # noqa E501
+                    context.rules_dict.update(control_imp_rules_dict)
+                    context.rules_params_dict.update(control_imp_rules_params_dict)
+                    ci_set_params = ControlInterface.get_set_params_from_item(control_imp)
+                    for imp_req in as_list(control_imp.implemented_requirements):
+                        control_part_id_map = part_id_map.get(imp_req.control_id, {})
+                        control_rules = ControlInterface.get_rule_list_for_item(imp_req)
+                        if control_rules:
+                            status = ControlInterface.get_status_from_props(imp_req)
+                            comp_info = ComponentImpInfo(imp_req.description, control_rules, status)
+                            self._catalog_interface.add_comp_info(imp_req.control_id, context.comp_name, '', comp_info)
+                            set_params = copy.deepcopy(ci_set_params)
+                            set_params.update(ControlInterface.get_set_params_from_item(imp_req))
+                            for set_param in set_params.values():
+                                self._catalog_interface.add_comp_set_param(
+                                    imp_req.control_id, context.comp_name, set_param
+                                )
+                            for statement in as_list(imp_req.statements):
+                                rule_list = ControlInterface.get_rule_list_for_item(statement)
+                                if rule_list:
+                                    status = ControlInterface.get_status_from_props(statement)
+                                    if statement.statement_id not in control_part_id_map:
+                                        label = statement.statement_id
+                                        logger.warning(
+                                            f'No statement label found for statement id {label}.  Defaulting to {label}.'  # noqa E501
+                                        )
+                                    else:
+                                        label = control_part_id_map[statement.statement_id]
+                                    comp_info = ComponentImpInfo(statement.description, rule_list, status)
+                                    self._catalog_interface.add_comp_info(
+                                        imp_req.control_id, context.comp_name, label, comp_info
+                                    )
+
+        # now have all rules in context.rules_dict and all rules_params in context.rules_params_dict
+        # all set-params per component for each control are in the cat interface
+        # all comp-infos by control and part are in the cat interface
+        #
+        # can now write out catalog and pull from the markdown:
+        # header for param values to set during assem
+        # prose and status for This System
+        # status for all parts that still have rules
+
+        for control_id, _ in self._catalog_interface._control_comp_dicts.items():
+            control_file_path = self._catalog_interface.get_control_file_path(context.md_root, control_id)
+            if not control_file_path:
+                logger.warning(f'Control {control_id} referenced by component is not in the resolved catalog.')
+                continue
+            control_file_path.parent.mkdir(exist_ok=True, parents=True)
+            control = self._catalog_interface.get_control(control_id)
+            _, group_title, _ = self._catalog_interface.get_group_info_by_control(control_id)
+            # merge the md_header and md_comp_dict with info in cat_interface for this control
+            self._catalog_interface._merge_header_and_comp_dict(control, control_file_path, context)
+            control_writer = ControlWriter()
+            control_writer.write_control_for_editing(
+                context, control, control_file_path.parent, group_title, part_id_map, []
+            )
+
+    def write_catalog_as_component_markdown(
+        self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]
+    ) -> None:
+        """Write out the catalog as component markdown."""
+
+        def _update_values(set_param: comp.SetParameter) -> None:
+            if set_param.param_id in control_param_dict:
+                control_param_dict[set_param.param_id].values = set_param.values
+            if set_param.param_id in new_context.rules_param_vals:
+                values = ', '.join([v.__root__ for v in as_list(set_param.values)])
+                new_context.rules_param_vals[set_param.param_id] = values
+
+        control_ids_in_comp_imp = [
+            imp_req.control_id for imp_req in as_list(context.control_implementation.implemented_requirements)
+        ]
+
+        for control in self._catalog_interface.get_all_controls_from_catalog(True):
+            if control.id in control_ids_in_comp_imp:
+                new_context = ControlContext.clone(context)
+                control_param_dict = ControlInterface.get_control_param_dict(control, False)
+                for set_param in as_list(new_context.control_implementation.set_parameters):
+                    _update_values(set_param)
+                for imp_req in as_list(new_context.control_implementation.implemented_requirements):
+                    if imp_req.control_id == control.id:
+                        for set_param in as_list(imp_req.set_parameters):
+                            _update_values(set_param)
+
+                if control_param_dict:
+                    new_context.yaml_header[const.PARAM_VALUES_TAG] = {}
+                    for key, param in control_param_dict.items():
+                        new_context.yaml_header[const.PARAM_VALUES_TAG][key] = none_if_empty(
+                            ControlInterface.param_to_str(param, ParameterRep.VALUE_OR_EMPTY_STRING)
+                        )
+
+                self._write_control_into_dir(new_context, control, part_id_map, [])
+
+    def write_catalog_as_catalog(self, context: ControlContext, part_id_map: Dict[str, Dict[str, str]]) -> None:
+        """Write the catalog as a simple catalog."""
+        # write out the controls
+        for control in self._catalog_interface.get_all_controls_from_catalog(True):
+            # here we do special handling of how set-parameters merge with the yaml header
+            new_context = ControlContext.clone(context)
+
+            control_param_dict = ControlInterface.get_control_param_dict(control, False)
+            set_param_dict: Dict[str, str] = {}
+            for param_id, param_dict in control_param_dict.items():
+                tmp_dict = ModelUtils.parameter_to_dict(param_dict, True)
+                values = tmp_dict.get('values', None)
+                new_dict = {'values': values}
+                set_param_dict[param_id] = new_dict
+            if set_param_dict:
+                if const.SET_PARAMS_TAG not in new_context.yaml_header:
+                    new_context.yaml_header[const.SET_PARAMS_TAG] = {}
+                if new_context.overwrite_header_values:
+                    # update the control params with new values
+                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
+                        if key in control_param_dict:
+                            set_param_dict[key] = value
+                else:
+                    # update the control params with any values in yaml header not set in control
+                    # need to maintain order in the set_param_dict
+                    for key, value in new_context.yaml_header[const.SET_PARAMS_TAG].items():
+                        if key in control_param_dict and key not in set_param_dict:
+                            set_param_dict[key] = value
+                new_context.yaml_header[const.SET_PARAMS_TAG] = set_param_dict
+            elif const.SET_PARAMS_TAG in new_context.yaml_header:
+                # need to cull any params that are not in control
+                pop_list: List[str] = []
+                for key in new_context.yaml_header[const.SET_PARAMS_TAG].keys():
+                    if key not in control_param_dict:
+                        pop_list.append(key)
+                for pop in pop_list:
+                    new_context.yaml_header[const.SET_PARAMS_TAG].pop(pop)
+
+            self._write_control_into_dir(new_context, control, part_id_map, [])
+
+    def _write_control_into_dir(
+        self,
+        context: ControlContext,
+        control: cat.Control,
+        part_id_map: Dict[str, Dict[str, str]],
+        found_control_alters: List[prof.Alter]
+    ):
+        # we need to create the dir structure on demand because we don't know a priori what groups are included
+        _, group_title, _ = self._catalog_interface.get_group_info_by_control(control.id)
+        group_dir = context.md_root
+        control_path = self._catalog_interface.get_control_path(control.id)
+        for sub_dir in control_path:
+            group_dir = group_dir / sub_dir
+            if not group_dir.exists():
+                group_dir.mkdir(parents=True, exist_ok=True)
+
+        writer = ControlWriter()
+        writer.write_control_for_editing(context, control, group_dir, group_title, part_id_map, found_control_alters)
