+++ trestle/core/catalog_validator.py
@@ -18,7 +18,7 @@ import logging
 
 from trestle.common.common_types import TopLevelOscalModel
 from trestle.common.list_utils import as_list
-from trestle.core.catalog_interface import CatalogInterface
+from trestle.core.catalog.catalog_interface import CatalogInterface
 from trestle.core.validator import Validator
 from trestle.oscal.catalog import Catalog
