+++ trestle/core/resolver/modify.py
@@ -22,7 +22,7 @@ from trestle.common.common_types import OBT
 from trestle.common.const import RESOLUTION_SOURCE, TRESTLE_INHERITED_PROPS_TRACKER
 from trestle.common.err import TrestleNotFoundError
 from trestle.common.list_utils import as_list, get_item_from_list, none_if_empty
-from trestle.core.catalog_interface import CatalogInterface
+from trestle.core.catalog.catalog_interface import CatalogInterface
 from trestle.core.control_interface import ParameterRep
 from trestle.core.pipeline import Pipeline
 from trestle.oscal import OSCAL_VERSION, common
