+++ trestle/tasks/xlsx_helper.py
@@ -26,7 +26,7 @@ from openpyxl.utils import get_column_letter
 from trestle import __version__
 from trestle.common.err import TrestleError
 from trestle.common.list_utils import is_ordered_sublist
-from trestle.core.catalog_interface import CatalogInterface
+from trestle.core.catalog.catalog_interface import CatalogInterface
 from trestle.oscal.catalog import Catalog
 from trestle.tasks.base_task import TaskBase
